LPS-70657 Move
